yum: don't require rpmdev-setuptree

+++ packages/yum/build-rpm.sh
@@ -88,10 +88,16 @@ if ! rpm -q mecab-devel > /dev/null; then
 base=http://download.fedoraproject.org/pub/fedora/linux/releases/15/Everything/source/SRPMS
 srpm=\$1
 
-rm -rf rpmbuild
+cat <<EOM > ~/.rpmmacros
+%_topdir \$HOME/rpmbuild
+EOM
 
-rm -rf .rpmmacros
-rpmdev-setuptree
+rm -rf rpmbuild
+mkdir -p rpmbuild/SOURCES
+mkdir -p rpmbuild/SPECS
+mkdir -p rpmbuild/BUILD
+mkdir -p rpmbuild/RPMS
+mkdir -p rpmbuild/SRPMS
 
 mkdir -p dependencies/RPMS
 mkdir -p dependencies/SRPMS
@@ -127,8 +133,20 @@ run yum clean ${yum_options} packages
 cat <<EOF > $BUILD_SCRIPT
 #!/bin/sh
 
-rm -rf .rpmmacros
-rpmdev-setuptree
+if [ -x /usr/bin/rpmdev-setuptree ]; then
+    rm -rf .rpmmacros
+    rpmdev-setuptree
+else
+    cat <<EOM > ~/.rpmmacros
+%_topdir \$HOME/rpmbuild
+EOM
+
+    mkdir -p rpmbuild/SOURCES
+    mkdir -p rpmbuild/SPECS
+    mkdir -p rpmbuild/BUILD
+    mkdir -p rpmbuild/RPMS
+    mkdir -p rpmbuild/SRPMS
+fi
 
 if test -f /tmp/${SOURCE_BASE_NAME}-$VERSION-*.src.rpm; then
     if ! rpm -Uvh /tmp/${SOURCE_BASE_NAME}-$VERSION-*.src.rpm; then
